# web/packages/fot-shared/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "FOT POS API",
    "version": "v2"
  },
  "paths": {
...
    "/api/settings/print": {
      "get": {
        "tags": [
          "FOT.Pos.Api"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrintSettingsDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "FOT.Pos.Api"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePrintSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/settings/cashboxes": {
      "get": {
        "tags": [
          "FOT.Pos.Api"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "FOT.Pos.Api"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePosCashBoxSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/settings/business-period": {
      "get": {
        "tags": [
          "FOT.Pos.Api"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "FOT.Pos.Api"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBusinessPeriodSettingsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PrintSettingsDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "headerText": {
            "type": "string",
            "nullable": true
          },
          "headerDescription": {
            "type": "string",
            "nullable": true
          },
          "footerText": {
            "type": "string",
            "nullable": true
          },
          "qrCodeText": {
            "type": "string",
            "nullable": true
          },
          "logoUrl": {
            "type": "string",
            "nullable": true
          },
          "showLogo": {
            "type": "boolean"
          },
          "showBarcode": {
            "type": "boolean"
          },
          "showArticleNumber": {
            "type": "boolean"
          },
          "copies": {
            "type": "integer",
            "format": "int32"
          },
          "paperWidthMm": {
            "type": "integer",
            "format": "int32"
          },
          "fontSize": {
            "type": "integer",
            "format": "int32"
          },
          "logoMaxHeightPx": {
            "type": "integer",
            "format": "int32"
          },
          "showQrCode": {
            "type": "boolean"
          },
          "showCashier": {
            "type": "boolean"
          },
          "showSalesman": {
            "type": "boolean"
          },
          "showDiscountDetails": {
            "type": "boolean"
          },
          "showItemTable": {
            "type": "boolean"
          },
          "showSubtotal": {
            "type": "boolean"
          },
          "showPaymentLines": {
            "type": "boolean"
          },
          "autoPrint": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateBusinessPeriodSettingsRequest": {
        "type": "object",
        "properties": {
          "weekStartDay": {
            "type": "integer",
            "format": "int32"
          },
          "weekLengthDays": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "UpdatePosCashBoxSettingsRequest": {
        "type": "object",
        "properties": {
          "qiMasterAccount": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "qiMasterAccountBank": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "giftMasterAccount": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "giftMasterAccountBank": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdatePrintSettingsRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "headerText": {
            "type": "string",
            "nullable": true
          },
          "headerDescription": {
            "type": "string",
            "nullable": true
          },
          "footerText": {
            "type": "string",
            "nullable": true
          },
          "qrCodeText": {
            "type": "string",
            "nullable": true
          },
          "showLogo": {
            "type": "boolean"
          },
          "showBarcode": {
            "type": "boolean"
          },
          "showArticleNumber": {
            "type": "boolean"
          },
          "copies": {
            "type": "integer",
            "format": "int32"
          },
          "paperWidthMm": {
            "type": "integer",
            "format": "int32"
          },
          "fontSize": {
            "type": "integer",
            "format": "int32"
          },
          "logoMaxHeightPx": {
            "type": "integer",
            "format": "int32"
          },
          "showQrCode": {
            "type": "boolean"
          },
          "showCashier": {
            "type": "boolean"
          },
          "showSalesman": {
            "type": "boolean"
          },
          "showDiscountDetails": {
            "type": "boolean"
          },
          "showItemTable": {
            "type": "boolean"
          },
          "showSubtotal": {
            "type": "boolean"
          },
          "showPaymentLines": {
            "type": "boolean"
          },
          "autoPrint": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Bearer token",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}